+++ src/dataclocklib/utility.py
@@ -23,7 +23,7 @@ Functions:
     add_colorbar: Add a colorbar to a figure, using the provided axis.
     add_text: Create annotation text on an Axes.
     assign_ring_wedge_columns: Assign ring & wedge columns to a DataFrame.
-    calculate_figure_dimensions: Calculate an optimal data clock figure size.
+    get_figure_dimensions: Calculate an optimal data clock figure size.
 
 Constants:
     VALID_STYLES: Valid font styles.
@@ -31,7 +31,7 @@ Constants:
 
 import math
 from collections import defaultdict
-from typing import Optional, Tuple, get_args
+from typing import Optional, Sequence, Tuple, get_args
 
 import numpy as np
 from matplotlib import colormaps
@@ -40,12 +40,13 @@ from matplotlib.cm import ScalarMappable
 from matplotlib.colorbar import Colorbar
 from matplotlib.colors import Normalize
 from matplotlib.figure import Figure
-from matplotlib.lines import Line2D
 from matplotlib.text import Text
-from numpy.typing import DTypeLike
-from pandas import DataFrame
+from numpy.typing import DTypeLike, NDArray
+from pandas import DataFrame, MultiIndex
+from pypalettes import load_cmap
 
-from dataclocklib.typing import CmapNames, FontStyle, Mode
+from dataclocklib.exceptions import ModeError
+from dataclocklib.typing import Aggregation, CmapNames, FontStyle, Mode
 
 VALID_STYLES: Tuple[FontStyle, ...] = get_args(FontStyle)
 
@@ -53,7 +54,8 @@ VALID_STYLES: Tuple[FontStyle, ...] = get_args(FontStyle)
 def add_colorbar(
     ax: Axes,
     fig: Figure,
-    cmap_name: CmapNames,
+    cmap_name: str,
+    cmap_reverse: bool,
     vmax: float,
     dtype: DTypeLike = np.float64,
 ) -> Colorbar:
@@ -72,7 +74,7 @@ def add_colorbar(
     """
     colorbar_ticks = np.linspace(1, vmax, 5, dtype=dtype)
 
-    cmap = colormaps.get_cmap(cmap_name)
+    cmap = load_cmap(cmap_name, cmap_type="continuous", reverse=cmap_reverse)
     cmap.set_under("w")
     cmap_norm = Normalize(1, vmax)
 
@@ -89,6 +91,68 @@ def add_colorbar(
 
     colorbar.ax.tick_params(direction="out")
     return colorbar
+
+
+def add_wedge_labels(
+    ax: Axes,
+    font_scale_factor: float,
+    ring_scale_factor: float,
+    ring_text_spacing: float,
+    max_radius: int,
+    theta: NDArray,
+    width: float,
+    wedge_labels: Sequence[str],
+) -> None:
+    """Add scaled and rotated labels around each data clock wedge.
+
+    Labels are placed using Axes.text to facilitate custom rotation
+    of the text, which is based on the angle of the wedge being
+    annotated. The text is scaled based on the size of the chart
+    Figure and padded away from the polar axis based on the number
+    of rings in the chart.
+
+    Args:
+        ax (Axes): Chart Axis.
+        font_scale_factor (float): Scale factor based on current figure size.
+        ring_scale_factor (float): Scale factor based on number of rings.
+        ring_text_spacing (float): Text label distance from polar axis.
+        max_radius (int): Maximum radius (unique rings + 1).
+        theta (NDArray): Angles (radians) for each data clock wedge.
+        width (float): Width of each wedge (2 * Pi / number of wedges).
+        wedge_labels (Sequence[str]): Label text for each wedge.
+
+        Returns:
+            None
+    """
+    if ring_scale_factor > 3:
+        ring_text_spacing = ring_text_spacing * (ring_scale_factor**0.61)
+    else:
+        ring_text_spacing = ring_text_spacing * ring_scale_factor
+
+    # place labels in the centre of each wedge
+    for idx, angle in enumerate(theta + width / 2):
+        # convert to degrees for text rotation
+        angle_deg = np.rad2deg(angle)
+
+        if (0 <= angle_deg < 90) or (270 <= angle_deg <= 360):
+            rotation = -angle_deg
+        else:
+            rotation = 180 - angle_deg
+
+        ax.text(
+            angle,
+            max_radius + ring_text_spacing,
+            wedge_labels[idx],
+            rotation=rotation,
+            rotation_mode="anchor",
+            transform=ax.transData,
+            family="sans-serif",
+            fontsize=11 * font_scale_factor,
+            weight="medium",
+            style="normal",
+            ha="center",
+            va="center",
+        )
 
 
 def add_text(
@@ -109,7 +173,72 @@ def add_text(
     return ax.text(x, y, s, **kwargs)
 
 
-def assign_ring_wedge_columns(
+def aggregate_temporal_columns(
+    data: DataFrame, agg_column: str, agg: Aggregation, mode: Mode
+) -> DataFrame:
+    """Aggregate values in agg_column using pass aggregate function.
+
+    Groups the DataFrame by the temporal 'ring' and 'wedge' columns,
+    before applying the aggregate function to the chosen aggregation
+    column.
+
+    NOTE: The 'ring' & 'wedge' columns are assigned by the utility function
+    assign_temporal_columns.
+
+    Args:
+        data (DataFrame): DataFrame containing data to aggregate.
+        agg_column (str): DataFrame Column to aggregate.
+        agg (Aggregation): Aggregation function; 'count', 'mean', 'median',
+            'mode' & 'sum'.
+        mode (Mode): A mode key representing the temporal bins used in the
+            chart; 'YEAR_MONTH', 'YEAR_WEEK', 'WEEK_DAY', 'DOW_HOUR' &
+            'DAY_HOUR'.
+
+    Raises:
+        ModeError: Unexpected mode value is passed.
+        ValueError: Missing 'ring' & 'wedge' columns.
+
+    Returns:
+        A DataFrame with aggregate values in a new column named after the
+        aggregate function.
+    """
+    columns = ["ring", "wedge"]
+    if not set(columns).issubset(data.columns):
+        raise ValueError(f"Expected DataFrame columns: {columns}")
+
+    unique_rings = data["ring"].unique()
+    match mode:
+        case "YEAR_MONTH":
+            unique_wedges = tuple(range(1, 13))
+        case "YEAR_WEEK":
+            unique_wedges = range(1, 53)
+        case "WEEK_DAY":
+            unique_wedges = range(0, 7)
+        case "DOW_HOUR":
+            unique_rings = range(0, 7)
+            unique_wedges = range(0, 24)
+        case "DAY_HOUR":
+            unique_wedges = range(0, 24)
+        case _:
+            raise ModeError(mode, get_args(Mode))
+
+    # groupby 'ring' & 'wedge' values and apply aggregate function agg
+    data_agg = data.groupby(columns, as_index=False)[agg_column].agg(agg)
+    data_agg = data_agg.set_axis([*columns, agg], axis="columns")
+
+    # index with all possible combinations of ring & wedge values
+    product_idx = MultiIndex.from_product(
+        [unique_rings, unique_wedges], names=columns
+    )
+
+    # populate any rows for missing ring/wedge combinations
+    data_agg = data_agg.set_index(columns).reindex(product_idx).reset_index()
+
+    # replace NaN values created for missing missing ring/wedge combinations
+    return data_agg.fillna(0)
+
+
+def assign_temporal_columns(
     data: DataFrame, date_column: str, mode: Mode
 ) -> DataFrame:
     """Assign ring & wedge columns to a DataFrame based on mode.
@@ -134,7 +263,7 @@ def assign_ring_wedge_columns(
     # year | January - December
     if mode == "YEAR_MONTH":
         mode_map[mode]["ring"] = data[date_column].dt.year
-        mode_map[mode]["wedge"] = data[date_column].dt.month_name()
+        mode_map[mode]["wedge"] = data[date_column].dt.month
     # year | weeks 1 - 52
     if mode == "YEAR_WEEK":
         mode_map[mode]["ring"] = data[date_column].dt.year
@@ -146,7 +275,7 @@ def assign_ring_wedge_columns(
         week = data[date_column].dt.isocalendar().week
         year = data[date_column].dt.year
         mode_map[mode]["ring"] = week + year * 100
-        mode_map[mode]["wedge"] = data[date_column].dt.day_name()
+        mode_map[mode]["wedge"] = data[date_column].dt.day_of_week
     # days 1 - 7 (Monday - Sunday) | 00:00 - 23:00
     if mode == "DOW_HOUR":
         mode_map[mode]["ring"] = data[date_column].dt.day_of_week
@@ -159,7 +288,7 @@ def assign_ring_wedge_columns(
     return data.assign(**mode_map[mode]).astype({"ring": "int64"})
 
 
-def calculate_figure_dimensions(wedges: int) -> tuple[float, float]:
+def get_figure_dimensions(wedges: int) -> tuple[float, float]:
     """Calculate an optimal data clock figure size based on wedge count.
 
     For most data clock charts, a minimum of 0.70 inches of figure space per
@@ -169,7 +298,6 @@ def calculate_figure_dimensions(wedges: int) -> tuple[float, float]:
     NOTE: The minimum figure size is capped at (10.0, 10.0).
 
     Example:
-      # 'DOW_HOUR' mode has 24 wedges for each of the 7 rings
       >>> calculate_figure_dimensions(168)
       (11, 11)